+++ 06.containers/source/mylist/mylist.h
@@ -52,10 +52,10 @@ public:
     T operator [] (int index) const override;  
     list &operator = (const list &other);
     
-    size_t size() const override { return m_size; }                          // return size of list (number of nodes)
+    int size() const override { return m_size; }                          // return size of list (number of nodes)
 
 private:
-    size_t m_size = 0;                                                       // size of list (number of nodes)
+    int m_size = 0;                                                       // size of list (number of nodes)
     Node<T> *m_begin = nullptr;                                              // pointer to the first node
     Node<T> *m_end = nullptr;                                                // pointer to the last node
 };
